services/calendar_service.py
import datetime
import logging
from typing import List, Dict, Any

# Assuming google-api-python-client is installed, but we handle imports defensively.
try:
    from google.oauth2.credentials import Credentials
    from googleapiclient.discovery import build
    from googleapiclient.errors import HttpError
except ImportError:
    pass

logger = logging.getLogger(__name__)

def fetch_upcoming_events(days: int = 7) -> List[Dict[str, Any]]:
    """
    Fetches events for the next 'days' days from the user's Google Calendar.
    Requires proper OAuth2 credentials setup which is assumed to be handled
    via environment variables or a token file.
    
    Returns a list of simplified event dictionaries.
    """
    try:
        # TODO: Implement robust credential loading here.
        # For now, this is a placeholder structure demonstrating the try/except blocks.
        # creds = Credentials.from_authorized_user_file('token.json', SCOPES)
        creds = None # Replace with actual credentials load
        
        if not creds:
            logger.warning("No Google Calendar credentials available. Returning empty events.")
            return []
            
        service = build('calendar', 'v3', credentials=creds)

        # Call the Calendar API
        now_dt = datetime.datetime.utcnow()
        timeMin = now_dt.isoformat() + 'Z'  # 'Z' indicates UTC time
        timeMax = (now_dt + datetime.timedelta(days=days)).isoformat() + 'Z'
        
        logger.info(f"Fetching calendar events from {timeMin} to {timeMax}")
        
        events_result = service.events().list(
            calendarId='primary', 
            timeMin=timeMin,
            timeMax=timeMax,
            maxResults=10, 
            singleEvents=True,
            orderBy='startTime'
        ).execute()
        
        events = events_result.get('items', [])
        
        parsed_events = []
        for event in events:
            start = event['start'].get('dateTime', event['start'].get('date'))
            parsed_events.append({
                "summary": event.get('summary', 'Untitled Event'),
                "start": start,
                "description": event.get('description', '')
            })
            
        return parsed_events

    except HttpError as error:
        logger.error(f"An HTTP error occurred interacting with Google Calendar API: {error}")
        return []
    except Exception as e:
        logger.error(f"An unexpected error occurred fetching calendar events: {e}")
        return []
