requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
openai>=1.0.0
langchain>=0.1.0
motor>=3.3.0
pydantic>=2.0.0
playwright>=1.38.0
apscheduler>=3.10.0
python-multipart>=0.0.6
requests>=2.31.0
beautifulsoup4>=4.12.0
google-api-python-client>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
langchain-openai

